+++ semgrep/tests/e2e/targets/basic/nosem.py
@@ -1,6 +1,22 @@
+# nosem: rules.test-nosem
+test_nosem_func()
+
+# NOSEM: rules.test-nosem
+test_nosem_func()
+
+# nosemgrep: rules.test-nosem
+test_nosem_func()
+
+# NOSEMGREP
+test_nosem_func()
+
+# nOseMgREP: rules.test-nosem
+test_nosem_func()
+
 test_nosem_func()  # nosem: rules.test-nosem
 test_nosem_func()  # NOSEM: rules.test-nosem
 test_nosem_func()  # nosemgrep: rules.test-nosem
 test_nosem_func()  # NOSEMGREP
 test_nosem_func()  # nOseMgREP: rules.test-nosem
+
 test_nosem_func()